Deleted unnesesary stuff
