+++ src/app/shared/components/menu/menu.component.html
@@ -1,0 +1,17 @@
+<div>
+    <mat-sidenav-container class="sidenav-container" class="Menu">
+        <mat-nav-list>
+            <a mat-list-item href="#"><mat-icon class="Icon">home</mat-icon>Home</a>
+            <a mat-list-item href="#"><mat-icon class="Icon">account_box</mat-icon>Profile</a>
+            <a mat-list-item href="#"><mat-icon class="Icon">description</mat-icon>Problems</a>
+            <a mat-list-item href="#"><mat-icon class="Icon">done_outline</mat-icon>Solved problems</a>
+            <a mat-list-item href="#"><mat-icon class="Icon">help</mat-icon>Help</a>
+            <a mat-list-item href="#"><mat-icon class="Icon">list</mat-icon>Solution queue</a>
+        </mat-nav-list>
+    </mat-sidenav-container>
+    <div class="Content">
+        <!-- write here -->
+    </div>
+</div>
+
+
