+++ src/app/shared/components/menu/menu.component.sass
@@ -1,0 +1,15 @@
+.Menu
+    left: 0
+    top: 0
+    position: fixed
+    width: 200px
+    height: 100%
+
+.Content
+    position: absolute
+    margin-left: 5px
+    left:205px
+
+.Icon
+    margin-right: 5px
+
