+++ src/app/shared/components/menu/menu.component.ts
@@ -1,0 +1,10 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'app-menu',
+    templateUrl: './menu.component.html',
+    styleUrls: ['./menu.component.sass'],
+})
+export class MenuComponent implements OnInit {
+    ngOnInit(): void {}
+}
