--- src/app/shared/menu/menu.component.html
@@ -1,32 +1,0 @@
-<mat-sidenav-container class="sidenav-container">
-    <mat-sidenav #drawer class="sidenav" fixedInViewport
-        [attr.role]="(isHandset$ | async) ? 'dialog' : 'navigation'"
-        [mode]="(isHandset$ | async) ? 'over' : 'side'"
-        [opened]="(isHandset$ | async) === false">
-      <mat-toolbar><strong>Goder</strong></mat-toolbar>
-      <mat-nav-list>
-        <a mat-list-item href="#"><mat-icon>home</mat-icon>Home</a>
-        <a mat-list-item href="#"><mat-icon>account_box</mat-icon>Profile</a>
-        <a mat-list-item href="#"><mat-icon>description</mat-icon>Problems</a>
-        <a mat-list-item href="#"><mat-icon>done_outline</mat-icon>Solved problems</a>
-        <a mat-list-item href="#"><mat-icon>help</mat-icon>Help</a>
-        <a mat-list-item href="#"><mat-icon>list</mat-icon>Solution queue</a>
-      </mat-nav-list>
-    </mat-sidenav>
-    <mat-sidenav-content>
-      <mat-toolbar>
-        <button
-          type="button"
-          aria-label="Toggle sidenav"
-          mat-icon-button
-          (click)="drawer.toggle()"
-          *ngIf="isHandset$ | async">
-          <mat-icon>menu</mat-icon>
-        </button>
-        <!-- Upper content -->
-      </mat-toolbar>
-      <!-- Content of page here -->
-    </mat-sidenav-content>
-
-  </mat-sidenav-container>
-
