--- src/app/shared/menu/menu.component.sass
@@ -1,20 +1,0 @@
-.sidenav-container 
-  height: 100%
-
-
-.sidenav 
-  width: 200px
-
-
-.sidenav 
-  .mat-toolbar 
-    background: inherit
-  
-  .mat-nav-list
-    a 
-      font-family: 'Lucida Sans'
-
-.mat-toolbar.mat-primary 
-  position: sticky
-  top: 0
-  z-index: 1
