--- src/app/shared/menu/menu.component.ts
@@ -1,19 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
-import { Observable } from 'rxjs';
-import { map, shareReplay } from 'rxjs/operators';
-
-@Component({
-    selector: 'app-menu',
-    templateUrl: './menu.component.html',
-    styleUrls: ['./menu.component.sass'],
-})
-export class MenuComponent implements OnInit {
-    isHandset$: Observable<boolean> = this.breakpointObserver.observe(Breakpoints.Handset).pipe(
-        map((result) => result.matches),
-        shareReplay()
-    );
-
-    constructor(private breakpointObserver: BreakpointObserver) {}
-    ngOnInit(): void {}
-}
